+++ cmd/metro2/main.go
@@ -17,9 +17,9 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/moov-io/metro2/pkg/file"
-	"github.com/moov-io/metro2/pkg/server"
-	"github.com/moov-io/metro2/pkg/utils"
+	"github.com/bloomcredit/moov-metro2/pkg/file"
+	"github.com/bloomcredit/moov-metro2/pkg/server"
+	"github.com/bloomcredit/moov-metro2/pkg/utils"
 )
 
 // utility functions