+++ pkg/lib/base_segment.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"unicode"
 
-	"github.com/moov-io/metro2/pkg/utils"
+	"github.com/bloomcredit/moov-metro2/pkg/utils"
 )
 
 var _ Record = (*BaseSegment)(nil)